Implemented contactor logic FSM
Starting from the requested state of the battery pack, it handles the transition logic to turn on the inverter rail or the charger rail.
At turn-on it precharge the output capacitor (through a dedicated resistor) in order to limit inrush currents and at turn-off it disharge it through another resistor.
The behaviour of the subsystem can be appreciated in contactor_logic_test.slx where both the controller and the contactor module are put together.